RHEL: Set vcpus hard limit to 240 for Power
Patchwork-id: 67655
O-Subject: [RHEL7.2 qemu-kvm-rhev PATCH] RHEL: Set vcpus hard limit to 240 for Power
Bugzilla: 1257781
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Thomas Huth <thuth@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>

On POWER, the kernel advertises a soft limit based on the number of CPU
threads on the host.  We want to allow exceeding this for testing purposes,
so we don't want to set hard limit to soft limit as on x86.

However the POWER hard limit advertised by the kernel is 2048 (== NR_CPUS)
but we only want to allow up to the number of vCPUs we actually test, so
we force the hard limit to 240.

Signed-off-by: David Gibson <dgibson@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>
(cherry picked from commit 8329392ab135fcaaa7fed94a6e8068dbed0ca473)
(cherry picked from commit 598c434cf95391ca5a3ddc0e955221e227bf3344)