block/vxhs: modularize VXHS via g_module
RH-Author: Jeffrey Cody <jcody@redhat.com>
Message-id: <8a91a423440b7a5a14e868279c772e99b865bfc6.1494281291.git.jcody@redhat.com>
Patchwork-id: 75046
O-Subject: [RHEV-7.4 qemu-kvm-rhev 4/4] block/vxhs: modularize VXHS via g_module
Bugzilla: 1265869
RH-Acked-by: Stefan Hajnoczi <stefanha@redhat.com>
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Miroslav Rezanina <mrezanin@redhat.com>

This converts the VXHS driver to be a runtime loaded module, with no
external build dependencies on libvxhs / libqnio.   This (libvxhs) is a
3rd party library, written by Veritas, to interface with the VXHS
server.

Red Hat is not going to distribute libvxhs, nor will Red Hat use libvxhs
in the build systems.  So that creates two criteria for the
modularization, for business reasons:

1. No runtime dependencies on libvxhs (aside from opening the library on
vxhs open)

2. No build dependencies on libvxhs packages.

There is support in QEMU for modular block drivers, however there are
two issues with using the built-in support:

    A. It is all-or-none; if modules are enabled all protocols are built
       as modules.  This wouldn't be that bad, as it would of course
       enable more granular dependencies for qemu rpm packages. But...

    B. It is not designed with criteria #2 in mind; it reduces runtime
       dependencies, not build dependencies.  The protocol libraries
       that are still built linked against external libraries and using
       external headers.

This patch uses g_module to load the libvxhs library, and incorporates
the libvxhs.h header in the build tree.  If block driver protocols are
also built as modules, libvxhs will still work and be built as a module
too, except the shared library will still not have any dependency on
libvxhs.

There are a few changes in this patch from upstream (aside from the
module loading aspects):

1. It enables VXHS support to be built as a protocl module if
--enable-modules is used during configure.

2. If the init call to iio_init() fails, populate errp with a
meaningful error message.

3. Since we are loading the library dynamically, make check the min and
max supported versions in the libvxhs library on load.

Patches for items #1 and #2 have been posted upstream.

It is expected that the libvxhs library is located at the following
pathname: /usr/lib64/qemu/libvxhs.so.1

VXHS support is only built for x86_64 in RHEV.

Signed-off-by: Jeff Cody <jcody@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>