blockdev: ignore aio=native for empty drives
RH-Author: John Snow <jsnow@redhat.com>
Message-id: <20170511205351.6337-2-jsnow@redhat.com>
Patchwork-id: 75070
O-Subject: [RHV-7.4 qemu-kvm-rhev PATCH 1/1] blockdev: ignore aio=native for empty drives
Bugzilla: 1402645
RH-Acked-by: Kevin Wolf <kwolf@redhat.com>
RH-Acked-by: Eric Blake <eblake@redhat.com>
RH-Acked-by: Max Reitz <mreitz@redhat.com>

This is a bit of a gross one; Upstream QEMU changed the way it handles
cache options with regards to removable media, associating options more
with the medium instead of the device. As part of that, it became
impossible to specify cache options on empty drives.

In the future, one would use blockdev-add instead to choose cache options
per-medium instead of per-device, but we're not there yet in libvirt so
we added a workaround downstream to simply ignore cache options on empty
CDROMs under the premise of "It actually never worked anyway."

As fallout from this decision, it is now no longer possible to specify
aio=native on empty CDROM devices either, as that requires the use of a
cache option that you can no longer specify. As a bad, gross, disgusting
workaround, simply ignore aio=native on empty drives until such time that
libvirt stops providing such configurations.

Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>