virtio_net: Bypass backends for MTU feature negotiation
RH-Author: Maxime Coquelin <maxime.coquelin@redhat.com>
Message-id: <20170601092415.2215-1-maxime.coquelin@redhat.com>
Patchwork-id: 75454
O-Subject: [RHEL7.4 qemu-kvm-rhev] virtio_net: Bypass backends for MTU feature negotiation
Bugzilla: 1452756
RH-Acked-by: Laurent Vivier <lvivier@redhat.com>
RH-Acked-by: Jens Freimann <jfreiman@redhat.com>
RH-Acked-by: Vlad Yasevich <vyasevic@redhat.com>
RH-Acked-by: Michael S. Tsirkin <mst@redhat.com>

This patch adds a new internal "x-mtu-bypass-backend" property
to bypass backends for MTU feature negotiation.

When this property is set, the MTU feature is negotiated as soon
as supported by the guest and a MTU value is set via the host_mtu
parameter. In case the backend advertises the feature (e.g. DPDK's
vhost-user backend), the feature negotiation is propagated down to
the backend.

When this property is not set, the backend has to support the MTU
feature for its negotiation to succeed.

For compatibility purpose, this property is disabled for machine
types RHEL7.3 (vs. v2.9 for upstream) and older.

Cc: Aaron Conole <aconole@redhat.com>
Suggested-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>
Reviewed-by: Vlad Yasevich <vyasevic@redhat.com>
Reviewed-by: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Michael S. Tsirkin <mst@redhat.com>
(cherry picked from commit 75ebec11afe49539f71cc1c494e3010f91c86adb)
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>